Switch from libsass to Dart Sass.
The latter is the flavor under current development by the Sass
maintainers.

It does not require C++ compilation, so this is more portable (eg. we
had issues compiling libsass under Windows with Bazel)

This currently uses the JavaScript transpiled version of Dart Sass. When
the Dart Bazel rules are caught up, we will also allow users to swap in
the native Dart version running in the Dart VM which is faster.